Refactor employee count types, improve filter logic, and update UI labels
- Change employee count from long to int in Java and MyBatis
- Update MyBatis XML to use int and improve null/empty filter checks
- Fix employee status value for "휴직 중" to "ON_LEAVE"
- Change pagination JS function and ensure size param is preserved
- Correct table alignment and reason variable in leave admin grants JSP
- Unify table cell alignment for consistency in leave admin grants view

+++ src/main/java/egovframework/let/attendance/repository/impl/EmployeeRepositoryImpl.java
@@ -31,7 +31,7 @@ public class EmployeeRepositoryImpl implements EmployeeRepositoryCustom {
 		int offset = (int) pageable.getOffset();
 		int limit = pageable.getPageSize();
 
-		long total = employeeDAO.countList(kw, dept, status);
+		int total = employeeDAO.countList(kw, dept, status);
 		List<Employee> content = total == 0 ? Collections.emptyList()
 				: employeeDAO.selectListPage(kw, dept, status, offset, limit);
